requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
et_xmlfile==2.0.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mutagen==1.47.0
openpyxl==3.1.5
pillow==12.1.1
Pygments==2.19.2
python-barcode==0.16.1
python-decouple==3.8
pytube==15.0.0
RapidFuzz==3.14.3
redis==7.3.0
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
setuptools==82.0.1
soupsieve==2.8.3
spotdl==3.9.6
spotipy==2.26.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
Unidecode==1.4.0
urllib3==2.6.3
whitenoise==6.12.0
yt-dlp==2026.3.13
ytmusicapi==1.11.5
